feat(types): all public api typed (#512)
